% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{compCdata_fun}
\alias{compCdata_fun}
\title{Construct the complete data including the observed entries and missing entries within a specifed window for control varieties}
\usage{
compCdata_fun(dat, w = Inf, cycle.list, type.prev = "cycle")
}
\arguments{
\item{dat}{a data frame with the following variable names: Variety, Cycle, DM_Trait}

\item{w}{window size for the missing entries}

\item{cycle.list}{a vector of cycle numbers spanned over in the CVEP data}

\item{type.prev}{a character, "cycle" or "history", to indicate what type of previous performance should be evaluated}
}
\value{
a data frame which includes Variety, Cycle, VarietybyCycle, DM_Trait, delta, New, DMtraitPrev
}
\description{
Construct the complete data including the observed entries and missing entries within a specifed window for control varieties
}
