% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{compRdata_fun}
\alias{compRdata_fun}
\title{Construct the complete data including the observed entries and missing entries within a specifed window for regular varieties}
\usage{
compRdata_fun(dat, w = Inf, year.list, type.prev = "year")
}
\arguments{
\item{dat}{a data frame with the following variable names: Variety, Year, DM_Trait}

\item{w}{window size for the missing entries}

\item{year.list}{a vector of year numbers spanned over in the CVEP data}

\item{type.prev}{a character, "year" or "history", to indicate what type of previous performance should be evaluated}
}
\value{
a data frame which includes Variety, Year, VarietybyYear, DM_Trait, delta, New, DMtraitPrev
}
\description{
Construct the complete data including the observed entries and missing entries within a specifed window for regular varieties
}
