% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{summary.fit}
\alias{summary.fit}
\title{Summarize the results from an object form \code{lmer}}
\usage{
\method{summary}{fit}(fit.lmer, effects.list, yes.IPW = T)
}
\arguments{
\item{fit.lmer}{a \code{\link{lmer.object}}}

\item{effects.list}{lists of year, location and variety}

\item{yes.IPW}{a logical varible indicating whether the model is fitted with weights}
}
\value{
a list including the estimates of fixed effects (includes the year effects and location effects) and random effects (including the variety, variety by year, and variety by location effects) as well as the estimates of the variance parameter (and standard deviation parameter) of random effects.
}
\description{
Summarize the results from an object form \code{lmer}
}
